feat(immutable-arraybuffer): freezable virtual typedarrays
