+++ packages/immutable-arraybuffer/src/immutable-arraybuffer-pony.js
@@ -1,253 +1,7 @@
-/* global globalThis */
-
-const {
-  ArrayBuffer,
-  Object,
-  Reflect,
-  Symbol,
-  TypeError,
-  Uint8Array,
-  WeakMap,
-  // Capture structuredClone before it can be scuttled.
-  structuredClone: optStructuredClone,
-  // eslint-disable-next-line no-restricted-globals
-} = globalThis;
-
-const { freeze, defineProperty, getPrototypeOf, getOwnPropertyDescriptor } =
-  Object;
-const { apply, ownKeys } = Reflect;
-const { toStringTag } = Symbol;
-
-const { prototype: arrayBufferPrototype } = ArrayBuffer;
-const { slice, transfer: optTransfer } = arrayBufferPrototype;
-// @ts-expect-error TS doesn't know it'll be there
-const { get: arrayBufferByteLength } = getOwnPropertyDescriptor(
-  arrayBufferPrototype,
-  'byteLength',
-);
-
-const typedArrayPrototype = getPrototypeOf(Uint8Array.prototype);
-const { set: uint8ArraySet } = typedArrayPrototype;
-// @ts-expect-error TS doesn't know it'll be there
-const { get: uint8ArrayBuffer } = getOwnPropertyDescriptor(
-  typedArrayPrototype,
-  'buffer',
-);
-
-/**
- * Copy a range of values from a genuine ArrayBuffer exotic object into a new
- * ArrayBuffer.
- *
- * @param {ArrayBuffer} realBuffer
- * @param {number} [start]
- * @param {number} [end]
- * @returns {ArrayBuffer}
- */
-const arrayBufferSlice = (realBuffer, start = undefined, end = undefined) =>
-  apply(slice, realBuffer, [start, end]);
-
-/**
- * Move the contents of a genuine ArrayBuffer exotic object into a new fresh
- * ArrayBuffer and detach the original source.
- * We can only do this on platforms that support `structuredClone` or
- * `ArrayBuffer.prototype.transfer`.
- * On other platforms, we can still emulate
- * `ArrayBuffer.prototoype.sliceToImmutable`, but not
- * `ArrayBuffer.prototype.transferToImmutable`.
- * Currently, these known-deficient platforms are
- * - Hermes
- * - Node.js <= 16
- * - Apparently some versions of JavaScriptCore that are still of concern.
- *
- * @param {ArrayBuffer} arrayBuffer
- * @returns {ArrayBuffer}
- */
-let optArrayBufferTransfer;
-
-if (optTransfer) {
-  optArrayBufferTransfer = arrayBuffer => apply(optTransfer, arrayBuffer, []);
-} else if (optStructuredClone) {
-  optArrayBufferTransfer = arrayBuffer => {
-    // Hopefully, a zero-length slice is cheap, but still enforces that
-    // `arrayBuffer` is a genuine `ArrayBuffer` exotic object.
-    arrayBufferSlice(arrayBuffer, 0, 0);
-    return optStructuredClone(arrayBuffer, {
-      transfer: [arrayBuffer],
-    });
-  };
-} else {
-  // Assignment is redundant, but remains for clarity.
-  optArrayBufferTransfer = undefined;
-}
-
-/**
- * If we could use classes with private fields everywhere, this would have
- * been a `this.#buffer` private field on an `ImmutableArrayBufferInternal`
- * class. But we cannot do so on Hermes. So, instead, we
- * emulate the `this.#buffer` private field, including its use as a brand check.
- * Maps from all and only emulated Immutable ArrayBuffers to real ArrayBuffers.
- *
- * @type {Pick<WeakMap<ArrayBuffer, ArrayBuffer>, 'get' | 'has' | 'set'>}
- */
-const buffers = new WeakMap();
-// Avoid post-hoc prototype lookups.
-for (const methodName of ['get', 'has', 'set']) {
-  defineProperty(buffers, methodName, { value: buffers[methodName] });
-}
-const getBuffer = immuAB => {
-  // Safe because this WeakMap owns its get method.
-  // eslint-disable-next-line @endo/no-polymorphic-call
-  const result = buffers.get(immuAB);
-  if (result) {
-    return result;
-  }
-  throw TypeError('Not an emulated Immutable ArrayBuffer');
-};
-
-// Omits `constructor` so `Array.prototype.constructor` is inherited.
-const ImmutableArrayBufferInternalPrototype = {
-  __proto__: arrayBufferPrototype,
-  get byteLength() {
-    return apply(arrayBufferByteLength, getBuffer(this), []);
-  },
-  get detached() {
-    getBuffer(this); // shim brand check
-    return false;
-  },
-  get maxByteLength() {
-    // Not underlying maxByteLength, which is irrelevant
-    return apply(arrayBufferByteLength, getBuffer(this), []);
-  },
-  get resizable() {
-    getBuffer(this); // shim brand check
-    return false;
-  },
-  get immutable() {
-    getBuffer(this); // shim brand check
-    return true;
-  },
-  slice(start = undefined, end = undefined) {
-    return arrayBufferSlice(getBuffer(this), start, end);
-  },
-  sliceToImmutable(start = undefined, end = undefined) {
-    // eslint-disable-next-line no-use-before-define
-    return sliceBufferToImmutable(getBuffer(this), start, end);
-  },
-  resize(_newByteLength = undefined) {
-    getBuffer(this); // shim brand check
-    throw TypeError('Cannot resize an immutable ArrayBuffer');
-  },
-  transfer(_newLength = undefined) {
-    getBuffer(this); // shim brand check
-    throw TypeError('Cannot detach an immutable ArrayBuffer');
-  },
-  transferToFixedLength(_newLength = undefined) {
-    getBuffer(this); // shim brand check
-    throw TypeError('Cannot detach an immutable ArrayBuffer');
-  },
-  transferToImmutable(_newLength = undefined) {
-    getBuffer(this); // shim brand check
-    throw TypeError('Cannot detach an immutable ArrayBuffer');
-  },
-  /**
-   * See https://github.com/endojs/endo/tree/master/packages/immutable-arraybuffer#purposeful-violation
-   */
-  [toStringTag]: 'ImmutableArrayBuffer',
-};
-
-// Better fidelity emulation of a class prototype
-for (const key of ownKeys(ImmutableArrayBufferInternalPrototype)) {
-  defineProperty(ImmutableArrayBufferInternalPrototype, key, {
-    enumerable: false,
-  });
-}
-
-/**
- * Emulates what would have been the encapsulated `ImmutableArrayBufferInternal`
- * class constructor. This function takes the `realBuffer` which its
- * result encapsulates. Security demands that this result has exclusive access
- * to the `realBuffer` it is given, which its callers must ensure.
- *
- * @param {ArrayBuffer} realBuffer
- * @returns {ArrayBuffer}
- */
-const makeImmutableArrayBufferInternal = realBuffer => {
-  const result = /** @type {ArrayBuffer} */ (
-    /** @type {unknown} */ ({
-      __proto__: ImmutableArrayBufferInternalPrototype,
-    })
-  );
-  // Safe because this WeakMap owns its set method.
-  // eslint-disable-next-line @endo/no-polymorphic-call
-  buffers.set(result, realBuffer);
-  return result;
-};
-// Since `makeImmutableArrayBufferInternal` MUST not escape,
-// this `freeze` is just belt-and-suspenders.
-freeze(makeImmutableArrayBufferInternal);
-
-/**
- * @param {ArrayBuffer} buffer
- * @returns {boolean}
- */
-// eslint-disable-next-line @endo/no-polymorphic-call
-export const isBufferImmutable = buffer => buffers.has(buffer);
-
-/**
- * Creates an immutable slice of the given buffer.
- * @param {ArrayBuffer} buffer The original buffer.
- * @param {number} [start] The start index.
- * @param {number} [end] The end index.
- * @returns {ArrayBuffer} The sliced immutable ArrayBuffer.
- */
-export const sliceBufferToImmutable = (
-  buffer,
-  start = undefined,
-  end = undefined,
-) => {
-  // Safe because this WeakMap owns its get method.
-  // eslint-disable-next-line @endo/no-polymorphic-call
-  let realBuffer = buffers.get(buffer);
-  if (realBuffer === undefined) {
-    realBuffer = buffer;
-  }
-  return makeImmutableArrayBufferInternal(
-    arrayBufferSlice(realBuffer, start, end),
-  );
-};
-
-let transferBufferToImmutable;
-if (optArrayBufferTransfer) {
-  /**
-   * Transfer the contents to a new Immutable ArrayBuffer
-   *
-   * @param {ArrayBuffer} buffer The original buffer.
-   * @param {number} [newLength] The start index.
-   * @returns {ArrayBuffer}
-   */
-  transferBufferToImmutable = (buffer, newLength = undefined) => {
-    if (newLength === undefined) {
-      buffer = optArrayBufferTransfer(buffer);
-    } else if (optTransfer) {
-      buffer = apply(optTransfer, buffer, [newLength]);
-    } else {
-      buffer = optArrayBufferTransfer(buffer);
-      const oldLength = buffer.byteLength;
-      // eslint-disable-next-line @endo/restrict-comparison-operands
-      if (newLength <= oldLength) {
-        buffer = arrayBufferSlice(buffer, 0, newLength);
-      } else {
-        const oldTA = new Uint8Array(buffer);
-        const newTA = new Uint8Array(newLength);
-        apply(uint8ArraySet, newTA, [oldTA]);
-        buffer = apply(uint8ArrayBuffer, newTA, []);
-      }
-    }
-    const result = makeImmutableArrayBufferInternal(buffer);
-    return /** @type {ArrayBuffer} */ (/** @type {unknown} */ (result));
-  };
-} else {
-  transferBufferToImmutable = undefined;
-}
-
-export const optTransferBufferToImmutable = transferBufferToImmutable;
+// Reexport just those exports that should be importable outside this package.
+// DO NOT reexport `hiddenBuffers`.
+export {
+  isBufferImmutable,
+  sliceBufferToImmutable,
+  optTransferBufferToImmutable,
+} from './immutable-arraybuffer-pony-internal.js';
